Fibonacci with array

+++ src/de/fhdw/prg1/hoh/set/arrays/Arrays.java
@@ -3,13 +3,13 @@ package de.fhdw.prg1.hoh.set.arrays;
 public class Arrays {
     public static void main(String[] args) {
         String result = "";
-        int[] myInts = new int[20];
+        long[] myInts = new long[20];
 
         for(int i = 0; i < myInts.length; i++){
             myInts[i] = i*2;
         }
 
-        for (int myInt : myInts) {
+        for (long myInt : myInts) {
             result += myInt + " ";
         }
         System.out.println(result);
